// src/types/collectionsTypes.ts
export enum Collections {
  CMS_MIGRATIONS = 'cmsMigrations',
  SCHEMAS = 'schemas',
  SCHEMA_FIELD_GROUPS = 'schemaFieldGroups',
  SCHEMA_FIELDS = 'schemaFields',

  ORGANIZATIONS = 'organizations',
  ORGANIZATION_USERS = 'organizationUsers',
  ORGANIZATION_USER_DEFINED_ROLS = 'organizationUserDefinedRols',

  LEADS = 'leads',
  USERS = 'users',

  USERS_ROLS = 'usersRols',

  PRODUCTS = 'products',
  STAFF = 'staff',
  USER_TOUCHPOINTS = 'userTouchpoints',
  ATTACHMENTS = 'attachments',

  USER_PRODUCTS = 'userProducts',
  OAUTH2_TOKENS = 'oauth2Tokens',
  USER_CALENDARS = 'userCalendars',
  USER_CALENDAR_EVENTS = 'userCalendarEvents',
  COMPANIES = 'companies',
  COMPANY_EMPLOYEES = 'companyEmployees',

  REMINDERS = 'reminders',

  CHAT_CONFIG = 'chatConfig',
  CHAT_MESSAGES = 'chatMessages',

  APP_CREATOR_WIDGETS = 'appCreatorWidgets',
  APP_CREATOR_FORMS = 'appCreatorForms',
  APP_CREATOR_COMPONENTS = 'appCreatorComponents',
  ENTITY_RELATIONSHIPS = 'entityRelationships',
}